Revert "Feature/horizontal scrolling"
