+++ bower.json
@@ -27,8 +27,7 @@
   ],
   "devDependencies": {
     "angular-mocks": "~1.3.0",
-    "web-component-tester": "*",
-    "Canteen": "https://github.com/platfora/Canteen.git"
+    "web-component-tester": "*"
   },
   "dependencies": {
     "angular": "~1.3.0",
@@ -43,7 +42,7 @@
     "webcomponentsjs": "^0.7.22",
     "widget-common": "https://github.com/Rise-Vision/widget-common.git#1.8.2",
     "widget-settings-ui-core": "https://github.com/Rise-Vision/widget-settings-ui-core.git#0.4.3",
-    "widget-settings-ui-components": "https://github.com/Rise-Vision/widget-settings-ui-components.git#3.1.3"
+    "widget-settings-ui-components": "https://github.com/Rise-Vision/widget-settings-ui-components.git#3.1.2"
   },
   "resolutions": {
     "angular": "~1.3.0",
