+++ src/widget/components/spreadsheet.jsx
@@ -612,10 +612,6 @@ const Spreadsheet = React.createClass({
     return this.getRowCount() > 1;
   },
 
-  showHeader: function() {
-    return params.spreadsheet.hasHeader && !(params.scroll.direction == "left");
-  },
-
   render: function() {
     if (this.state.data) {
       this.setColumnCount();
@@ -623,7 +619,7 @@ const Spreadsheet = React.createClass({
 
       return(
         <div id="table">
-        {this.showHeader() ?
+        {params.spreadsheet.hasHeader ?
           <TableHeaderContainer
             align={params.format.header.fontStyle.align}
             data={this.getHeaders()}
@@ -643,7 +639,6 @@ const Spreadsheet = React.createClass({
               rowHeight={params.format.rowHeight}
               width={params.width}
               height={params.spreadsheet.hasHeader ? params.height - params.format.rowHeight : params.height}
-              bodyFontStyle={params.format.body.fontStyle}
               columnFormats={this.getColumnFormats()} />
           : false}
         </div>
