--- test/unit/widget/vertical.jsx
@@ -1,65 +1,0 @@
-import React from "react";
-import { shallow } from "enzyme";
-import { expect } from "chai";
-import TestUtils from "react-addons-test-utils";
-import Vertical from "../../../src/widget/components/vertical";
-import Table from "../../../src/widget/components/table";
-
-describe("<Vertical />", function() {
-  const onDone = function () {},
-    scroll = {},
-    align = "center",
-    className = "body_font-style",
-    width = 600,
-    height = 50,
-    totalCols = 3,
-    rowHeight = 50,
-    data = [["I am the walrus!", "1", "3"],
-      ["John is dead!", "500", "32"]],
-    columnFormats = [{ "width": 100 }, { "width": 200 }, { "width": 300 }];
-
-  var wrapper, renderedComponent;
-
-  beforeEach(function () {
-    var component = <Vertical onDone={onDone} scroll={scroll} data={data} align={align} class={className} totalCols={totalCols} rowHeight={rowHeight} width={width} height={height} columnFormats={columnFormats} />;
-
-    wrapper = shallow(component);
-    renderedComponent = TestUtils.renderIntoDocument(component);
-
-  });
-
-  describe("Page", function() {
-    it("Should set the page class to the section element", function() {
-      expect(renderedComponent.refs.page.className).to.equal("page");
-    });
-  });
-
-
-  describe("<Table />", function() {
-    it("Should contain a Table component", function() {
-      expect(wrapper.find(Table)).to.have.length(1);
-    });
-
-    it("Should have data prop", function() {
-      var expected = [["I am the walrus!", "1", "3"],
-                           ["John is dead!", "500", "32"]];
-      expect(wrapper.find(Table).props().data).to.deep.equal(expected);
-    });
-
-    it("Should have columnFormats prop", function() {
-      expect(wrapper.find(Table).props().columnFormats).to.deep.equal(columnFormats);
-    });
-
-    it("Should have width prop", function() {
-      expect(wrapper.find(Table).props().width).to.equal(width);
-    });
-
-    it("Should have height prop", function() {
-      expect(wrapper.find(Table).props().height).to.equal(height);
-    });
-
-    it("Should have totalCols prop", function() {
-      expect(wrapper.find(Table).props().totalCols).to.equal(totalCols);
-    });
-  });
-});
